Move user role change logic into switch component
